Allow files bigger than 2G to be deployed using ant tasks.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1693461 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/ant/AbstractCatalinaTask.java
@@ -185,7 +185,7 @@ public abstract class AbstractCatalinaTask extends BaseRedirectorHelperTask {
      * @exception BuildException if an error occurs
      */
     public void execute(String command, InputStream istream,
-                        String contentType, int contentLength)
+                        String contentType, long contentLength)
         throws BuildException {
 
         URLConnection conn = null;
